requirements.txt
azure-core==1.29.5
azure-identity==1.14.1
azure-storage-blob==12.18.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
idna==3.4
isodate==0.6.1
loguru==0.7.2
msal==1.24.1
msal-extensions==1.0.0
numpy==1.26.1
Office365-REST-Python-Client==2.5.1
pandas==2.1.1
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
